Revert "Bump to cocoa sdk v8.46.0 (#4103)"
This reverts commit 690a1697
