+++ modules/sentry-cocoa.properties
@@ -1,2 +1,2 @@
-version = 8.46.0
+version = 8.39.0
 repo = https://github.com/getsentry/sentry-cocoa
