+++ src/Sentry.Bindings.Cocoa/StructsAndEnums.cs
@@ -66,3 +66,25 @@ internal enum SentrySpanStatus : ulong
     OutOfRange,
     DataLoss
 }
+
+[Native]
+internal enum SentryLevel : ulong
+{
+    None = 0,
+    Debug = 1,
+    Info = 2,
+    Warning = 3,
+    Error = 4,
+    Fatal = 5
+}
+
+[Native]
+internal enum SentryTransactionNameSource : long
+{
+    Custom = 0,
+    Url = 1,
+    Route = 2,
+    View = 3,
+    Component = 4,
+    Task = 5
+}
